Allow customisation of which fact is used for hostnames
Previously, this role used the host part of ansible_fqdn. This should be the
same as ansible_hostname. In some cases we may wish to use the full FQDN, or
the nodename (ansible_nodename).

This change adds support for modifying the fact used via a
cluster_hostname_fact variable, which defaults to ansible_hostname.